Add MathApplet
MathApplet that brings up a new window and asks a math question. If the answer typed in is correct when the "Check" button is clicked it will display "Correct!" if is is incorrect then it will tell you the answer. "New problem" button randomly generates a new problem to ask the user.